+++ project/instance.go
@@ -79,28 +79,9 @@ func serviceToInstance(c *client.Client, p *types.Project, serviceName string, o
 		errs = errors.Join(errs, err)
 	}
 
-	var (
-		uid uint64
-		gid uint64
-	)
 	// User override - https://github.com/compose-spec/compose-spec/blob/main/05-services.md#user
-	if service.User != "" {
-		split := strings.Split(service.User, ":")
-
-		uid, err = strconv.ParseUint(split[0], 10, 64)
-		if err != nil {
-			return nil, nil, fmt.Errorf("cannot convert service '%v' user '%v' to int: %w", service.Name, split[0], err)
-		}
-		config["oci.uid"] = split[0]
-		if len(split) > 1 {
-			gid, err = strconv.ParseUint(split[1], 10, 64)
-			if err != nil {
-				return nil, nil, fmt.Errorf("cannot convert service '%v' user '%v' to int: %w", service.Name, split[1], err)
-			}
-		}
-	}
-
-	volumes, files, volumeResources, err := instanceVolumeDevices(c, p, service, image, uid, gid)
+	// A name resolves against the image, which is not pulled yet here.
+	volumes, files, volumeResources, err := instanceVolumeDevices(c, p, service, image, service.User)
 	if err != nil {
 		errs = errors.Join(errs, err)
 	}
@@ -145,8 +126,7 @@ func serviceToInstance(c *client.Client, p *types.Project, serviceName string, o
 		Dependencies:  instanceDependencyWaits(p, service, options),
 		Entrypoint:    service.Entrypoint,
 		Command:       service.Command,
-		UID:           uid,
-		GID:           gid,
+		User:          service.User,
 	}
 
 	ir, err := c.Resource(client.KindInstance, instanceName, instCfg)
@@ -620,7 +600,7 @@ func instanceProxyDevices(c *client.Client, devices []client.InstanceDevice, ser
 // instanceVolumeDevices builds disk, bind, and tmpfs devices for a service's
 // volumes plus the shm_size tmpfs. It returns any storage volume resources
 // and the files map for single-file binds.
-func instanceVolumeDevices(c *client.Client, p *types.Project, service types.ServiceConfig, image client.Resource, uid, gid uint64) ([]client.InstanceDevice, []client.InstanceFile, []client.Resource, error) {
+func instanceVolumeDevices(c *client.Client, p *types.Project, service types.ServiceConfig, image client.Resource, user string) ([]client.InstanceDevice, []client.InstanceFile, []client.Resource, error) {
 	var errs error
 	devices := []client.InstanceDevice{}
 	resources := []client.Resource{}
@@ -688,8 +668,7 @@ func instanceVolumeDevices(c *client.Client, p *types.Project, service types.Ser
 			volConfig := &client.StorageVolumeConfig{
 				Shifted:       shifted,
 				ImageResource: image,
-				UID:           uid,
-				GID:           gid,
+				User:          user,
 				Pool:          pool,
 				Prefetch:      prefetch,
 				Extensions:    extensions,
@@ -744,8 +723,7 @@ func instanceVolumeDevices(c *client.Client, p *types.Project, service types.Ser
 					volConfig := &client.StorageVolumeConfig{
 						Shifted:       shifted,
 						ImageResource: image,
-						UID:           uid,
-						GID:           gid,
+						User:          user,
 						HostPath:      cVol.Source,
 						Pool:          c.Config().DefaultStoragePool,
 					}